Moved smb validation functions to app_validation_checks.py and save_upgrade_script_config into upgrade_functions.py
